+++ src/infra/userRoles/updateUserRoles.ts
@@ -1,4 +1,5 @@
 import UserRole from '../../models/userRolesModel';
+import Role from '../../models/rolesModel'; // Modelo para validar roles
 import { Transaction } from 'sequelize';
 import sequelize from '../../config/database';
 
@@ -7,6 +8,16 @@ export class UpdateUserRolesSql {
     const transaction: Transaction = await sequelize.transaction();
 
     try {
+      // Validar que los IDs de roles existan
+      const validRoles = await Role.findAll({
+        where: { id: roleIds },
+        attributes: ['id'],
+      });
+
+      if (validRoles.length !== roleIds.length) {
+        throw new Error('Uno o más IDs de roles no son válidos.');
+      }
+
       // Eliminar los roles existentes del usuario
       await UserRole.destroy({
         where: { userId },
